test: cleanup + isolate on tempo mod
