+++ crates/common/Cargo.toml
@@ -92,6 +92,7 @@ chrono.workspace = true
 vergen = { workspace = true, features = ["build", "emit_and_set"] }
 
 [dev-dependencies]
+foundry-evm-hardforks.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
 axum = { workspace = true }
 tempfile.workspace = true
