+++ crates/common/src/tempo/tests.rs
@@ -1,0 +1,56 @@
+use alloy_provider::{Provider, ProviderBuilder};
+use eyre::WrapErr;
+use foundry_evm_hardforks::TempoHardfork;
+use serde::Deserialize;
+use std::env;
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct TempoForkSchedule {
+    schedule: Vec<ForkInfo>,
+    #[allow(dead_code)]
+    active: String,
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct ForkInfo {
+    name: String,
+}
+
+async fn check_fork_schedule(rpc_url: &str) -> eyre::Result<()> {
+    let provider = ProviderBuilder::new().connect_http(rpc_url.parse()?);
+    let schedule: TempoForkSchedule =
+        provider.raw_request("tempo_forkSchedule".into(), ()).await?;
+    for fork in &schedule.schedule {
+        fork.name.parse::<TempoHardfork>()?;
+    }
+    Ok(())
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_fork_schedule_parses_configured_rpcs() -> eyre::Result<()> {
+    let mut checked_any = false;
+
+    for (network, env_key) in [
+        ("mainnet", "TEMPO_MAINNET_RPC_URL"),
+        ("testnet", "TEMPO_TESTNET_RPC_URL"),
+        ("devnet", "TEMPO_DEVNET_RPC_URL"),
+    ] {
+        let rpc_url = match env::var(env_key) {
+            Ok(url) if !url.is_empty() => url,
+            _ => continue,
+        };
+        checked_any = true;
+
+        check_fork_schedule(&rpc_url)
+            .await
+            .wrap_err_with(|| format!("[{network}] {env_key}={rpc_url}"))?;
+    }
+
+    if !checked_any {
+        eprintln!("Missing Tempo RPC env vars. Skipping Tempo fork schedule compatibility test.");
+    }
+
+    Ok(())
+}
